Title: 두 수의 합, Time: 3092ms, Memory: 34MB, Status: Time Limit Exceed - Codetree

+++ 250608/두 수의 합/sum-of-two-num.java
@@ -3,7 +3,7 @@ public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         int n = sc.nextInt();
-        int k = sc.nextInt();
+        long k = sc.nextLong();
         int[] arr = new int[n];
         for (int i = 0; i < n; i++) {
             arr[i] = sc.nextInt();
